chore: rearrange UI for language and appearance
