+++ packages/shared/locales/en.json
@@ -14,9 +14,9 @@
             "body": "Help the developers improve Firefly by automatically sending diagnostic data when an error or crash occurs.",
             "checkbox": "Send crash reports to the IOTA Foundation"
         },
-        "appearance": {
-            "title": "Appearance",
-            "body": "Choose your app theme."
+        "languageAndAppearance": {
+            "title": "Language & appearance",
+            "body": "Choose your language and theme"
         },
         "profile": {
             "title": "Create your {protocol} profile",
