+++ packages/shared/routes/dashboard/index.js
@@ -1,0 +1,5 @@
+export { default as Settings } from './settings/Settings.svelte'
+export { default as Staking } from './staking/Staking.svelte'
+export { default as Wallet } from './wallet/Wallet.svelte'
+
+export { default as Dashboard } from './Dashboard.svelte'
