+++ packages/shared/routes/index.js
@@ -1,32 +1,5 @@
-// splash
+export * from './dashboard'
+export * from './login'
+export * from './onboarding'
+
 export { default as Splash } from 'shared/routes/Splash.svelte'
-// onboarding/setup
-export { default as Welcome } from 'shared/routes/setup/Welcome.svelte'
-export { default as Legal } from 'shared/routes/setup/legal/Legal.svelte'
-export { default as CrashReporting } from 'shared/routes/setup/CrashReporting.svelte'
-export { default as Appearance } from 'shared/routes/setup/Appearance.svelte'
-export { default as Secure } from 'shared/routes/setup/Secure.svelte'
-export { default as Profile } from 'shared/routes/setup/Profile.svelte'
-export { default as Setup } from 'shared/routes/setup/Setup.svelte'
-export { default as Create } from 'shared/routes/setup/Create.svelte'
-export { default as OnboardingProtocol } from 'shared/routes/setup/OnboardingProtocol.svelte'
-export { default as OnboardingNetwork } from 'shared/routes/setup/OnboardingNetwork.svelte'
-export { default as OnboardingCustomNetwork } from 'shared/routes/setup/OnboardingCustomNetwork.svelte'
-export { default as Ledger } from 'shared/routes/setup/ledger/Ledger.svelte'
-export { default as Password } from 'shared/routes/setup/Password.svelte'
-export { default as ProtectRouter } from 'shared/routes/setup/protect/ProtectRouter.svelte'
-export { default as Backup } from 'shared/routes/setup/backup/Backup.svelte'
-export { default as Import } from 'shared/routes/setup/import/Import.svelte'
-export { default as Balance } from 'shared/routes/setup/Balance.svelte'
-export { default as ClaimRewards } from 'shared/routes/setup/ClaimRewards.svelte'
-export { default as Migrate } from 'shared/routes/setup/migrate/Migrate.svelte'
-export { default as Congratulations } from 'shared/routes/setup/Congratulations.svelte'
-
-// dashboard
-export { default as Dashboard } from 'shared/routes/dashboard/Dashboard.svelte'
-export { default as Wallet } from 'shared/routes/dashboard/wallet/Wallet.svelte'
-export { default as Settings } from 'shared/routes/dashboard/settings/Settings.svelte'
-export { default as Staking } from 'shared/routes/dashboard/staking/Staking.svelte'
-
-// login
-export { default as LoginRouter } from 'shared/routes/login/LoginRouter.svelte'
