+++ packages/shared/routes/login/index.js
@@ -1,0 +1,1 @@
+export { default as LoginRouter } from './LoginRouter.svelte'
