+++ packages/shared/routes/onboarding/Balance.svelte
