+++ packages/shared/routes/onboarding/OnboardingNetwork.svelte
