+++ packages/shared/routes/onboarding/backup/views/RecoveryPhrase.svelte
