+++ packages/shared/routes/onboarding/backup/views/VerifyRecoveryPhrase.svelte
