+++ packages/shared/routes/onboarding/index.js
@@ -1,0 +1,21 @@
+export { default as Backup } from './backup/Backup.svelte'
+export { default as Import } from './import/Import.svelte'
+export { default as Ledger } from './ledger/Ledger.svelte'
+export { default as Legal } from './legal/Legal.svelte'
+export { default as Migrate } from './migrate/Migrate.svelte'
+export { default as ProtectRouter } from './protect/ProtectRouter.svelte'
+
+export { default as Balance } from './Balance.svelte'
+export { default as ClaimRewards } from './ClaimRewards.svelte'
+export { default as Congratulations } from './Congratulations.svelte'
+export { default as CrashReporting } from './CrashReporting.svelte'
+export { default as Create } from './Create.svelte'
+export { default as LanguageAndAppearanceView } from './LanguageAndAppearanceView.svelte'
+export { default as OnboardingCustomNetwork } from './OnboardingCustomNetwork.svelte'
+export { default as OnboardingNetwork } from './OnboardingNetwork.svelte'
+export { default as OnboardingProtocol } from './OnboardingProtocol.svelte'
+export { default as Password } from './Password.svelte'
+export { default as Profile } from './Profile.svelte'
+export { default as Secure } from './Secure.svelte'
+export { default as Setup } from './Setup.svelte'
+export { default as Welcome } from './Welcome.svelte'
