+++ packages/shared/routes/onboarding/ledger/views/AccountIndex.svelte
