+++ packages/shared/routes/onboarding/ledger/views/GenerateNewAddress.svelte
