+++ packages/shared/routes/onboarding/ledger/views/InstallationGuide.svelte
