+++ packages/shared/routes/onboarding/ledger/views/LegacyIntro.svelte
