+++ packages/shared/routes/onboarding/ledger/views/SwitchApps.svelte
