+++ packages/shared/routes/onboarding/legal/Content.svelte
