+++ packages/shared/routes/onboarding/legal/Legal.svelte
