+++ packages/shared/routes/onboarding/migrate/views/BundleMiningWarning.svelte
