+++ packages/shared/routes/onboarding/migrate/views/SecuringSpentAddresses.svelte
