+++ packages/shared/routes/onboarding/protect/ProtectRouter.svelte
