+++ packages/shared/routes/onboarding/protect/views/ChooseProtectionMethodView.svelte
