+++ packages/shared/routes/onboarding/protect/views/SetupPinProtectionView.svelte
