+++ packages/shared/routes/onboarding/protect/views/index.js
